RES: Don't search std/core crates in workspace in 2018 Edition
Our implementation already processes std/core crates so with lookup into workspace
it leads to multiple resolution results and prevents from success resolution of all items in std/core crates
