+++ src/test/kotlin/org/rust/lang/core/resolve/RsStdlibResolveTest.kt
@@ -7,10 +7,20 @@ package org.rust.lang.core.resolve
 
 import org.rust.ProjectDescriptor
 import org.rust.WithStdlibRustProjectDescriptor
+import org.rust.cargo.project.model.cargoProjects
+import org.rust.cargo.project.workspace.CargoWorkspace
 import org.rust.lang.core.types.infer.TypeInferenceMarks
 
 @ProjectDescriptor(WithStdlibRustProjectDescriptor::class)
 class RsStdlibResolveTest : RsResolveTestBase() {
+
+    override fun runTest() {
+        for (edition in CargoWorkspace.Edition.values()) {
+            project.cargoProjects.setEdition(edition)
+            super.runTest()
+        }
+    }
+
     fun `test resolve fs`() = stubOnlyResolve("""
     //- main.rs
         use std::fs::File;
@@ -480,18 +490,7 @@ class RsStdlibResolveTest : RsResolveTestBase() {
         }
     """, TypeInferenceMarks.questionOperator)
 
-    fun `test try! macro`() = checkByCode("""
-        struct S { field: u32 }
-                    //X
-        fn foo() -> Result<S, ()> { unimplemented!() }
 
-        //noinspection RsTryMacro
-        fn main() {
-            let s = try!(foo());
-            s.field;
-            //^
-        }
-    """)
 
     fun `test try! macro with aliased Result`() = checkByCode("""
         mod io {
@@ -542,18 +541,6 @@ class RsStdlibResolveTest : RsResolveTestBase() {
 
         fn foo(v: Vec) {}
                  //^ unresolved
-    """)
-
-    fun `test resolve with no_std attribute 2`() = stubOnlyResolve("""
-    //- main.rs
-        #![no_std]
-
-        extern crate alloc;
-
-        use alloc::vec::Vec;
-
-        fn foo(v: Vec) {}
-                 //^ .../liballoc/vec.rs
     """)
 
     fun `test inherent impl have higher priority than derived`() = checkByCode("""
